move to from generic utils to dbt.project module
